# Backend/api/openapi/teams-notification-api.yaml
openapi: 3.0.3
info:
  title: Teams Notification API (Go implementation)
  version: 1.0.0
  description: |
    This specification documents the REST endpoints implemented in the Teams Notification Go service.
    Includes all available routes for internal and external APIs.

    ## API Structure
    - **External APIs** (`/api/v1`): Public APIs for external clients
    - **Internal APIs** (`/internal/v1`): Internal management APIs
    - **System APIs**: Health, metrics, and configuration endpoints

    ## Recent Updates (v1.0.0)
    - **Unified Configuration**: Environment-based configuration system
    - **Error Handling**: Standardized error responses with proper HTTP status codes
    - **Token Management**: Redis-backed token caching for improved performance
    - **Monitoring**: Enhanced health checks and system metrics
    - **Logging**: Structured logging with configurable levels and formats
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.teamsnotify.com
    description: Production server
paths:
...
  /api/v1/notify:
    post:
      summary: Send notification (External API)
      tags: [External API]
      description: Send a notification to Teams using project notify key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExternalNotifyRequest"
      responses:
        "200":
          description: Notification sent successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExternalNotifyResponse"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "404":
          $ref: "#/components/responses/NotFoundError"
  /api/v1/destinations/{notifyKey}:
    get:
      summary: Get project destinations (External API)
      tags: [External API]
      description: Get destinations for a project using notify key
      parameters:
        - name: notifyKey
          in: path
          required: true
          schema:
            type: string
          description: Project notify key
      responses:
        "200":
          description: Project destinations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectDestinationsResponse"
        "404":
          $ref: "#/components/responses/NotFoundError"
...
  /api/v1/messages:
    post:
      summary: Handle Bot Framework messages
      tags: [External API]
      description: Handle incoming Bot Framework Activity messages
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BotFrameworkActivity"
      responses:
        "200":
          description: Message processed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /api/v1/messages/test:
    post:
      summary: Test proactive message
      tags: [External API]
      description: Send a test proactive message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProactiveTestRequest"
      responses:
        "200":
          description: Test message sent successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # Internal API endpoints (/internal/v1)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BadRequestError:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFoundError:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
        details:
          type: string
          description: Detailed error information
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
...
    ExternalNotifyRequest:
      type: object
      required:
        - notifyKey
      properties:
        notifyKey:
          type: string
          description: Project notify key
          example: "my-project-key"
        message:
          type: string
          description: Notification message
          example: "Hello from external API"
        messageType:
          type: string
          enum: [text, file, adaptive_card]
          default: text
          description: Type of message
        priority:
          type: string
          enum: [low, normal, high]
          default: normal
          description: Message priority
        targets:
          type: array
          items:
            type: string
          description: List of target emails, conversation IDs, or 'all'
          example: ["all"]
        mentions:
          type: array
          items:
            type: string
          description: List of user mentions
          example: ["@user1", "@user2"]
        metadata:
          type: object
          additionalProperties: true
          description: Additional metadata
          example: { "source": "external", "version": "1.0" }
        attachments:
          type: array
          description: Bot Framework attachments array. For Adaptive Card, set contentType to application/vnd.microsoft.card.adaptive and put card JSON in content.
          items:
            type: object
            properties:
              contentType:
                type: string
                example: application/vnd.microsoft.card.adaptive
              contentUrl:
                type: string
                nullable: true
              name:
                type: string
                nullable: true
              thumbnailUrl:
                type: string
                nullable: true
              content:
                type: object
                additionalProperties: true
          example:
            - contentType: application/vnd.microsoft.card.adaptive
              content:
                $schema: "http://adaptivecards.io/schemas/adaptive-card.json"
                type: AdaptiveCard
                version: "1.3"
                body:
                  - type: TextBlock
                    text: "Hello from Adaptive Card"
                    wrap: true
        templateId:
          type: string
          format: uuid
          description: UUID of the template to use
        templateData:
          type: object
          additionalProperties: true
          description: Key-value pairs for template variable substitution
        isTemplate:
          type: boolean
          description: Flag to indicate if this is a template based message
    ExternalNotifyResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/NotifyResponse"
        error:
          type: string
    NotifyResponse:
      type: object
      properties:
        notificationId:
          type: string
          format: uuid
        status:
          type: string
        message:
          type: string
        destinationsCount:
          type: integer
        estimated_delivery_time:
          type: string

    # Company schemas
...
    BotFrameworkActivity:
      type: object
      properties:
        type:
          type: string
        id:
          type: string
        timestamp:
          type: string
          format: date-time
        channelId:
          type: string
        from:
          $ref: "#/components/schemas/ChannelAccount"
        conversation:
          $ref: "#/components/schemas/ConversationAccount"
        recipient:
          $ref: "#/components/schemas/ChannelAccount"
        text:
          type: string
        attachments:
          type: array
          items:
            $ref: "#/components/schemas/Attachment"
    ChannelAccount:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
    ConversationAccount:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
    Attachment:
      type: object
      properties:
        contentType:
          type: string
        contentUrl:
          type: string
        name:
          type: string
    MessageResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
    ProactiveTestRequest:
      type: object
      required:
        - conversationId
        - message
      properties:
        conversationId:
          type: string
          description: Teams conversation ID
        message:
          type: string
          description: Test message content

    # Provision schemas
...
    ProjectDestinationsResponse:
      type: object
      properties:
        notifyKey:
          type: string
        destinations:
          type: array
          items:
            $ref: "#/components/schemas/DestinationResponse"
    DestinationResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [personal, group, channel]
        conversationId:
          type: string
        displayName:
          type: string
        tenantId:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time

    # Configuration schemas
...